refactor: Renames analytics.integrations.github.github To analytics.integrations.github.client
